+++ CoffeeMachine.Code/Models/CatalogRecord.cs
@@ -1,3 +1,3 @@
 namespace CoffeeMachine.Code.Models;
 
-public record CatalogRecord(DrinkType DrinkType, string DrinkCode, double Price);
+public record CatalogRecord(DrinkType DrinkType, string DrinkCode, decimal Price);
